sc-meta cli SenderArgs refactor
